requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
graphql-core==3.2.3
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
strawberry-django-auth==0.374.0
strawberry-django-plus==1.35.2
strawberry-graphql==0.155.2
strawberry-graphql-django==0.9
typing_extensions==4.4.0
